+++ CHANGES.rst
@@ -9,6 +9,7 @@ Unreleased
 - Fix `model_to_dict` to detect changes in a parent model when using
   `inherit=True` (backwards-incompatible for users who were directly
   using previous version) (gh-576)
+- Use an iterator for `clean_duplicate_history`
 
 2.7.3 (2019-07-15)
 ------------------